SetiaModule/requirements.txt
google-cloud-storage
google-cloud-aiplatform
vertexai
functions-framework
pandas
ipython